fix rendering offset, works a tiny bit better

+++ lib/binaries/unpack.dart
@@ -3,5 +3,5 @@ import 'dart:typed_data';
 
 void main(String args) {
   final result = new File(args[0]).readAsBytesSync();
-  print(new Uint16List.fromList(result));
+  print(new Uint8List.fromList(result));
 }
